Module aws_sdk_codegurureviewer::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

not_found_exception

See NotFoundException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AssociateRepositoryError

ConflictException

The requested operation would cause a conflict with the current state of a service

CreateCodeReviewError

DescribeCodeReviewError

DescribeRecommendationFeedbackError

DescribeRepositoryAssociationError

DisassociateRepositoryError

InternalServerException

The server encountered an internal error and is unable to complete the request.

ListCodeReviewsError

ListRecommendationFeedbackError

ListRecommendationsError

ListRepositoryAssociationsError

ListTagsForResourceError

NotFoundException

The resource specified in the request was not found.

PutRecommendationFeedbackError

ResourceNotFoundException

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

ValidationException

The input fails to satisfy the specified constraints.

Enums

AssociateRepositoryErrorKind

CreateCodeReviewErrorKind

DescribeCodeReviewErrorKind

DescribeRecommendationFeedbackErrorKind

DescribeRepositoryAssociationErrorKind

DisassociateRepositoryErrorKind

ListCodeReviewsErrorKind

ListRecommendationFeedbackErrorKind

ListRecommendationsErrorKind

ListRepositoryAssociationsErrorKind

ListTagsForResourceErrorKind

PutRecommendationFeedbackErrorKind

TagResourceErrorKind

UntagResourceErrorKind