Module aws_sdk_codegurureviewer::model

Modules

branch_diff_source_code_type

See BranchDiffSourceCodeType

code_artifacts

See CodeArtifacts

code_commit_repository

See CodeCommitRepository

code_review

See CodeReview

code_review_summary

See CodeReviewSummary

code_review_type

See CodeReviewType

commit_diff_source_code_type

See CommitDiffSourceCodeType

event_info

See EventInfo

kms_key_details

See KmsKeyDetails

metrics

See Metrics

metrics_summary

See MetricsSummary

recommendation_feedback

See RecommendationFeedback

recommendation_feedback_summary

See RecommendationFeedbackSummary

recommendation_summary

See RecommendationSummary

repository

See Repository

repository_analysis

See RepositoryAnalysis

repository_association

See RepositoryAssociation

repository_association_summary

See RepositoryAssociationSummary

repository_head_source_code_type

See RepositoryHeadSourceCodeType

request_metadata

See RequestMetadata

rule_metadata

See RuleMetadata

s3_bucket_repository

See S3BucketRepository

s3_repository

See S3Repository

s3_repository_details

See S3RepositoryDetails

source_code_type

See SourceCodeType

third_party_source_repository

See ThirdPartySourceRepository

Structs

BranchDiffSourceCodeType

CodeArtifacts

Code artifacts are source code artifacts and build artifacts used in a repository analysis or a pull request review.

CodeCommitRepository

Information about an Amazon Web Services CodeCommit repository. The CodeCommit repository must be in the same

CodeReview

CodeReviewSummary

CodeReviewType

CommitDiffSourceCodeType

EventInfo

Information about an event. The event might be a push, pull request, scheduled request, or another type of event.

KmsKeyDetails

An object that contains:

Metrics

MetricsSummary

RecommendationFeedback

RecommendationFeedbackSummary

RecommendationSummary

Repository

RepositoryAnalysis

A code review type that analyzes all code under a specified branch in an associated

RepositoryAssociation

Information about a repository association. The

RepositoryAssociationSummary

Summary information about a repository association. The

RepositoryHeadSourceCodeType

RequestMetadata

Metadata that is associated with a code review. This applies to both pull request

RuleMetadata

Metadata about a rule. Rule metadata includes an ID, a name, a list of tags, and a

S3BucketRepository

S3Repository

S3RepositoryDetails

SourceCodeType

ThirdPartySourceRepository

Enums

AnalysisType

EncryptionOption

JobState

ProviderType

Reaction

RecommendationCategory

RepositoryAssociationState

Severity

Type

VendorName