Struct aws_sdk_codegurureviewer::model::code_review_summary::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CodeReviewSummary

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the code review.

pub fn set_name(self, input: Option<String>) -> Self

pub fn code_review_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the CodeReview object.

pub fn set_code_review_arn(self, input: Option<String>) -> Self

pub fn repository_name(self, input: impl Into<String>) -> Self

The name of the repository.

pub fn set_repository_name(self, input: Option<String>) -> Self

pub fn owner(self, input: impl Into<String>) -> Self

The owner of the repository. For an Amazon Web Services CodeCommit repository, this is the Amazon Web Services account ID of the account that owns the repository. For a GitHub, GitHub Enterprise Server, or Bitbucket repository, this is the username for the account that owns the repository. For an S3 repository, it can be the username or Amazon Web Services account ID.

pub fn set_owner(self, input: Option<String>) -> Self

pub fn provider_type(self, input: ProviderType) -> Self

The provider type of the repository association.

pub fn set_provider_type(self, input: Option<ProviderType>) -> Self

pub fn state(self, input: JobState) -> Self

The state of the code review.

The valid code review states are:

Completed: The code review is complete.

Pending: The code review started and has not completed or failed.

Failed: The code review failed.

Deleting: The code review is being deleted.

pub fn set_state(self, input: Option<JobState>) -> Self

pub fn created_time_stamp(self, input: Instant) -> Self

The time, in milliseconds since the epoch, when the code review was created.

pub fn set_created_time_stamp(self, input: Option<Instant>) -> Self

pub fn last_updated_time_stamp(self, input: Instant) -> Self

The time, in milliseconds since the epoch, when the code review was last updated.

pub fn set_last_updated_time_stamp(self, input: Option<Instant>) -> Self

pub fn type(self, input: Type) -> Self

The type of the code review.

pub fn set_type(self, input: Option<Type>) -> Self

pub fn pull_request_id(self, input: impl Into<String>) -> Self

The pull request ID for the code review.

pub fn set_pull_request_id(self, input: Option<String>) -> Self

pub fn metrics_summary(self, input: MetricsSummary) -> Self

The statistics from the code review.

pub fn set_metrics_summary(self, input: Option<MetricsSummary>) -> Self

pub fn source_code_type(self, input: SourceCodeType) -> Self

Specifies the source code that is analyzed in a code review.

pub fn set_source_code_type(self, input: Option<SourceCodeType>) -> Self

pub fn build(self) -> CodeReviewSummary

Consumes the builder and constructs a CodeReviewSummary